# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Produtos",
        "description": "Documentação da API de Produtos usando Swagger(antes de criar os produtos a(s) categoria precisam ser criada primeiro)",
        "contact": {
            "email": "suporte@exemplo.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/api/category": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Listar todas as categorias",
                "description": "Retorna a lista completa de categorias",
                "operationId": "getCategories",
                "responses": {
                    "200": {
                        "description": "Lista de categorias retornada com sucesso"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Categorias"
                ],
                "summary": "nova categoria",
                "description": "Criação da categoria",
                "operationId": "Criar uma nova Categoria",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Roupas"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mostra a categoria criada com sucesso"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/category/{category}": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "detalhes da categoria",
                "description": "Detalhamento da categoria",
                "operationId": "recuperar o categoria por id",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "ID da category (id e do tipo uuid precisa copiar o codigo id gerado na lista de categorias entre as aspas duplas e colar no campo abaixo)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostra os detalhes da categoria com sucesso"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Categorias"
                ],
                "summary": "atualizar categoria",
                "description": "Atualização da categoria",
                "operationId": "Atualizar uma categoria",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "ID da category (id e do tipo uuid precisa copiar o codigo id gerado na lista de categorias entre as aspas duplas e colar no campo abaixo)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Roupas quentes"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mostra atualização da categoria com sucesso"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categorias"
                ],
                "summary": "deletar categoria",
                "description": "categoria deletada",
                "operationId": "Deletar uma Categoria",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "ID da category (id e do tipo uuid precisa copiar o codigo id gerado na lista de categorias entre as aspas duplas e colar no campo abaixo)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostra a categoria deletada com sucesso"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "logar usúario",
                "description": "Login do usúario",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "2i2ww@google.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "T@rt@14#"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mostra o usuário logado com sucesso"
                    }
                }
            }
        },
        "/api/logout/{user}": {
            "post": {
                "tags": [
                    "Logout"
                ],
                "summary": "deslogar usúario",
                "description": "Logout do usúario",
                "operationId": "logout",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID do usúario (id e do tipo uuid precisa copiar o codigo id gerado na lista de produtos entre as aspas duplas e colar no campo abaixo)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Mostra o usuário deslogado com sucesso"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}